requirements.txt
audioread==3.0.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
decorator==5.2.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
moviepy==2.2.1
msgpack==1.1.1
numba==0.62.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
proglog==0.1.12
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
